+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 01.04.2018
+
+Facelift
+
+### Added
+- Samsung Internet, Yandex and Electron browsers support
+
+### Changed
+- Browser images served from [@browser-logos](https://github.com/alrra/browser-logos) now [#9](https://github.com/godban/browsers-support-badges/issues/9)
+
 ## [1.0.0] - 26.03.2018
 
 Re-Init